requirements.txt
waybackpy
requests
beautifulsoup4
python-dotenv
PyGithub
